+++ src/main/java/net/igneo/icv/event/ModEvents.java
@@ -1,10 +1,13 @@
 package net.igneo.icv.event;
 
 import net.igneo.icv.ICV;
+import net.igneo.icv.enchantment.AcrobaticEnchantment;
 import net.igneo.icv.enchantmentActions.PlayerEnchantmentActions;
 import net.igneo.icv.enchantmentActions.PlayerEnchantmentActionsProvider;
 import net.igneo.icv.networking.ModMessages;
 import net.igneo.icv.networking.packet.BlitzNBTUpdateS2CPacket;
+import net.igneo.icv.particle.ModParticles;
+import net.igneo.icv.sound.ModSounds;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
@@ -23,6 +26,7 @@ import net.minecraftforge.fml.LogicalSide;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.loading.FMLEnvironment;
 
+import static java.lang.Math.abs;
 import static net.igneo.icv.enchantment.BlitzEnchantment.ATTACK_SPEED_MODIFIER_UUID;
 
 @Mod.EventBusSubscriber(modid = ICV.MOD_ID)
@@ -56,6 +60,10 @@ public class ModEvents {
     public static void onPlayerTick(TickEvent.PlayerTickEvent event) {
         event.player.getCapability(PlayerEnchantmentActionsProvider.PLAYER_ENCHANTMENT_ACTIONS).ifPresent(enchVar -> {
 
+            if (enchVar.getAcrobatBonus() && (event.player.onGround() || event.player.isInFluidType() || event.player.isPassenger())) {
+                enchVar.setAcrobatBonus(false);
+            }
+
             //Blitz check
             if (enchVar.getBlitzBoostCount() > 0) {
                 if (System.currentTimeMillis() >= enchVar.getBlitzTime() + 1000) {
@@ -76,12 +84,22 @@ public class ModEvents {
 
             //Phantom pain check
             if (enchVar.getPhantomVictim() != null) {
-                if (System.currentTimeMillis() >= enchVar.getPhantomDelay() + 4000) {
-                    enchVar.getPhantomVictim().heal(enchVar.getPhantomHurt());
-                    enchVar.resetPhantomHurt();
-                    enchVar.deletePhantomVictim();
-                    System.out.println("healing previous entity");
+                if (enchVar.getPhantomVictim().isAlive()) {
+                    if (System.currentTimeMillis() >= enchVar.getPhantomDelay() + 4000) {
+                        ServerPlayer player = (ServerPlayer) event.player;
+                        ServerLevel level = player.serverLevel();
+                        level.sendParticles(player, ModParticles.PHANTOM_HEAL_PARTICLE.get(), true, enchVar.getPhantomVictim().getX(), enchVar.getPhantomVictim().getY() + 1.5, enchVar.getPhantomVictim().getZ(), 10, Math.random(), Math.random(), Math.random(), 0.5);
+                        level.playSound(null, enchVar.getPhantomVictim().blockPosition(), ModSounds.PHANTOM_HEAL.get(), SoundSource.PLAYERS, 0.25F, (float) 0.3 + (float) abs(Math.random() + 0.5));
+                        enchVar.getPhantomVictim().heal(enchVar.getPhantomHurt());
+                        enchVar.resetPhantomHurt();
+                        enchVar.deletePhantomVictim();
+                        System.out.println("healing previous entity");
+                    }
                 }
+            }
+
+            if (FMLEnvironment.dist.isClient()) {
+                AcrobaticEnchantment.onClientTick();
             }
         });
     }
